added basic validations

+++ backend/db/migrations/20240227060819-create-event.js
@@ -41,7 +41,7 @@ module.exports = {
                 type: Sequelize.INTEGER,
             },
             price: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.DECIMAL(6, 2),
             },
             startDate: {
                 type: Sequelize.DATE,
